Update linker plain linker file
Tell the linker that the zero page memory is matched with the zpage
memory space qualifier.

Also exclude the stack area from the memories and rely on having the
stack set up by TinyCore. We do not define a stack at all. If this
area is provided the linker will incorrectly think it can place things
in this 0100-01ff area which is highly unsafe.